frontend/requirements.txt
ursus-ssg==1.4.6
pyphen==0.17.2
holidays